backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
dotenv
python-dotenv
bcrypt==3.2.2
passlib
PyJWT
pydantic